Add convenience methods to StackAllocator and use it for the window title.
The StackAllocator can be used in the following way:

`char *title = (char *)m_StackAllocator->Allocate(sizeof(char) * 50);`

Will allocate `sizeof(char) * 50` bytes on the stack.

`SomeObject *obj = m_StackAllocator->make_new<SomeObject>(constructor_arguments_here);`

Will allocate room for a `SomeObject` on the stack and return a pointer to it.
If the object you want to allocate has a different number of arguments (or need references or `const`
something), add the signature to `StackAllocator.hpp`.

The stack allocator will be cleared at the beginning of each frame.

The stack allocator cannot delete objects, but only clear the entire stack until a specified point.
This is done by a call to `m_StackAllocator->FreeToMarker(marker)`.
A marker to a stack point can be obtained via `m_StackAllocator->GetMarker()`.

The stack allocator has 200 MB allocated at the moment.